project/requirements.txt
flask>=3.0.0
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
